refactor(port): share strict tree drift comparison
