+++ scripts/tests/test_pi.py
@@ -8,10 +8,10 @@ from pathlib import Path
 import pytest
 
 from scripts import build_pi_skills
-from scripts.build_codex_skills import _differences
 from scripts.port_lib import SOURCE_PLUGIN_DIR, TARGETS_DIR
 from scripts.port_lib import pi
 from scripts.port_lib.frontmatter import parse_file
+from scripts.port_lib.generated_tree import tree_differences
 
 
 def _tree_hash(root: Path) -> str:
@@ -166,7 +166,7 @@ def test_determinism_rollback_and_read_only_drift_detection(tmp_path, monkeypatc
     second = tmp_path / "second"
     pi.build(plugin, first)
     pi.build(plugin, second)
-    assert _differences(first, second) == []
+    assert tree_differences(first, second) == []
     assert _tree_hash(first) == _tree_hash(second)
 
     before = _tree_hash(first)
